use vscode storage for fql playground (#22)
ENG-5510
The fql playgournd file will now use vscode storage instead of the
users' project.  If the extension is active within a project, it will
use vscode project storage.  This means that for each different project
users' would have different playgrounds with different content.
If not in a project it will use vscode global storage.